+++ src/pandabear/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.4"
+__version__ = "0.3.2"
 
 
 # Set default logging handler to avoid "No handler found" warnings.
@@ -7,7 +7,7 @@ from logging import NullHandler
 
 logging.getLogger(__name__).addHandler(NullHandler())
 
-from pandabear.decorators import check_types
+from pandabear.decorators import check, check_types
 from pandabear.model import DataFrameModel, SeriesModel
 from pandabear.model_components import Field
-from pandabear.typing import DataFrame, Series
+from pandabear.typehints import DataFrame, Series